コントローラー名の変更
